chore: move `computeAuthWitHash` to aztec.js utils
