+++ yarn-project/aztec.js/src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './l1_contracts.js';
 export * from './l2_contracts.js';
 export * from './abi_types.js';
 export * from './cheat_codes.js';
+export * from './authwit.js';
