+++ yarn-project/end-to-end/src/e2e_cross_chain_messaging.test.ts
@@ -1,4 +1,4 @@
-import { AccountWallet, AztecAddress } from '@aztec/aztec.js';
+import { AccountWallet, AztecAddress, computeAuthWitHash } from '@aztec/aztec.js';
 import { Fr } from '@aztec/circuits.js';
 import { EthAddress } from '@aztec/foundation/eth-address';
 import { DebugLogger } from '@aztec/foundation/log';
@@ -6,7 +6,7 @@ import { TokenBridgeContract, TokenContract } from '@aztec/noir-contracts/types'
 import { TxStatus } from '@aztec/types';
 
 import { CrossChainTestHarness } from './fixtures/cross_chain_test_harness.js';
-import { computeAuthWitHash, delay, setup } from './fixtures/utils.js';
+import { delay, setup } from './fixtures/utils.js';
 
 describe('e2e_cross_chain_messaging', () => {
   let logger: DebugLogger;
