+++ yarn-project/end-to-end/src/e2e_public_cross_chain_messaging.test.ts
@@ -1,12 +1,12 @@
-import { AccountWallet, AztecAddress } from '@aztec/aztec.js';
-import { Fr, FunctionSelector } from '@aztec/circuits.js';
+import { AccountWallet, AztecAddress, computeAuthWitHash } from '@aztec/aztec.js';
+import { Fr } from '@aztec/circuits.js';
 import { EthAddress } from '@aztec/foundation/eth-address';
 import { DebugLogger } from '@aztec/foundation/log';
 import { TokenBridgeContract, TokenContract } from '@aztec/noir-contracts/types';
 import { TxStatus } from '@aztec/types';
 
 import { CrossChainTestHarness } from './fixtures/cross_chain_test_harness.js';
-import { delay, hashPayload, setup } from './fixtures/utils.js';
+import { delay, setup } from './fixtures/utils.js';
 
 describe('e2e_public_cross_chain_messaging', () => {
   let logger: DebugLogger;
@@ -94,14 +94,10 @@ describe('e2e_public_cross_chain_messaging', () => {
     // 4. Give approval to bridge to burn owner's funds:
     const withdrawAmount = 9n;
     const nonce = Fr.random();
-    const burnMessageHash = await hashPayload([
-      l2Bridge.address.toField(),
-      l2Token.address.toField(),
-      FunctionSelector.fromSignature('burn_public((Field),Field,Field)').toField(),
-      ownerAddress.toField(),
-      new Fr(withdrawAmount),
-      nonce,
-    ]);
+    const burnMessageHash = await computeAuthWitHash(
+      l2Bridge.address,
+      l2Token.methods.burn_public(ownerAddress, withdrawAmount, nonce).request(),
+    );
     await ownerWallet.setPublicAuth(burnMessageHash, true).send().wait();
 
     // 5. Withdraw owner's funds from L2 to L1
