+++ yarn-project/end-to-end/src/fixtures/utils.ts
@@ -11,8 +11,6 @@ import {
   createPXEClient,
   getSandboxAccountsWallets,
 } from '@aztec/aztec.js';
-import { CircuitsWasm, GeneratorIndex } from '@aztec/circuits.js';
-import { pedersenPlookupCompressWithHashIndex } from '@aztec/circuits.js/barretenberg';
 import {
   DeployL1Contracts,
   L1ContractArtifactsForDeployment,
@@ -44,11 +42,9 @@ import { NonNativeTokenContract, TokenBridgeContract, TokenContract } from '@azt
 import { PXEService, createPXEService, getPXEServiceConfig } from '@aztec/pxe';
 import {
   AztecNode,
-  FunctionCall,
   L2BlockL2Logs,
   LogType,
   PXE,
-  PackedArguments,
   TxStatus,
   UnencryptedL2Log,
   createAztecNodeRpcClient,
@@ -552,32 +548,4 @@ export const expectUnencryptedLogsFromLastBlockToBe = async (pxe: PXE, logMessag
   const asciiLogs = unrolledLogs.map(log => log.data.toString('ascii'));
 
   expect(asciiLogs).toStrictEqual(logMessages);
-};
-
-/**
- * Hash a payload to generate a signature on an account contract
- * @param payload - payload to hash
- * @returns the hashed message
- */
-export const hashPayload = async (payload: Fr[]) => {
-  return pedersenPlookupCompressWithHashIndex(
-    await CircuitsWasm.get(),
-    payload.map(fr => fr.toBuffer()),
-    GeneratorIndex.SIGNATURE_PAYLOAD,
-  );
-};
-
-/**
- * Compute an authentication witness message hash from a caller and a request
- * @param caller - The caller approved to make the call
- * @param request - The request to be made (function call)
- * @returns The message hash for the witness
- */
-export const computeAuthWitHash = async (caller: AztecAddress, request: FunctionCall) => {
-  return await hashPayload([
-    caller.toField(),
-    request.to.toField(),
-    request.functionData.selector.toField(),
-    (await PackedArguments.fromArgs(request.args, await CircuitsWasm.get())).hash,
-  ]);
 };
